core/cli: extract shared `patchProgram`
All of `--watch`, `--build`, and `--build-watch` need the same approach
to patching the `TS.Program`, so extract it from `perform-watch` into a
dedicated utility module.
